Rename NativeApi -> NativeAPI.
