+++ ScriptEngine/API/APIHelp.cpp
@@ -24,7 +24,7 @@
 #include <Engine/EngineOwnData.h>
 #include "APIHelp.h"
 
-#include "NativeApi.h"
+#include "NativeAPI.h"
 
 using namespace std;
