+++ ScriptEngine/API/NativePatch.cpp
@@ -1,4 +1,4 @@
-#include "NativeApi.h"
+#include "NativeAPI.h"
 
 #include <third-party/ModUtils/ModUtils.h>
 #include <Utils/TypeConversionHelper.hpp>
